+++ earth2observe/chirps.py
@@ -1,34 +1,47 @@
 import datetime as dt
 import os
 from ftplib import FTP
+from pathlib import Path
 
 import numpy as np
 import pandas as pd
 from joblib import Parallel, delayed
 from osgeo import gdal
 from pyramids.raster import Raster
+from pyramids.utils import extractFromGZ
 
-from earth2observe.utils import extractFromGZ, print_progress_bar
+from earth2observe.abstractdatasource import AbstractCatalog, AbstractDataSource
+from earth2observe.utils import print_progress_bar
 
 
-class CHIRPS:
+class CHIRPS(AbstractDataSource):
     """CHIRPS."""
+
+    api_url: str = "data.chc.ucsb.edu"
+    start_date: str = "1981-01-01"
+    end_date: str = "Now"
+    temporal_resolution = ["daily", "monthly"]
+    lat_bondaries = [-50, 50]
+    lon_boundaries = [-180, 180]
+    globe_fname = "chirps-v2.0"
+    clipped_fname = "P_CHIRPS.v2.0"
 
     def __init__(
         self,
-        start: str = "",
-        end: str = "",
-        lat_lim: list = [],
-        lon_lim: list = [],
-        time: str = "daily",
+        temporal_resolution: str = "daily",
+        start: str = None,
+        end: str = None,
         path: str = "",
+        variables: list = None,
+        lat_lim: list = None,
+        lon_lim: list = None,
         fmt: str = "%Y-%m-%d",
     ):
         """CHIRPS.
 
         Parameters
         ----------
-        time (str, optional):
+        temporal_resolution (str, optional):
             'daily' or 'monthly'. Defaults to 'daily'.
         start (str, optional):
             [description]. Defaults to ''.
@@ -45,56 +58,105 @@ class CHIRPS:
         fmt (str, optional):
             [description]. Defaults to "%Y-%m-%d".
         """
-        # Define timestep for the timedates
-        self.lat_lim = []
-        self.lon_lim = []
-        if time.lower() == "daily":
-            self.time_freq = "D"
-            self.output_folder = os.path.join(path, "Precipitation", "CHIRPS", "Daily")
-        elif time.lower() == "monthly":
-            self.time_freq = "MS"
-            self.output_folder = os.path.join(
-                path, "Precipitation", "CHIRPS", "Monthly"
-            )
-        else:
-            raise KeyError("The input time interval is not supported")
-        self.time = time
+        super().__init__(
+            start=start,
+            end=end,
+            variables=variables,
+            temporal_resolution=temporal_resolution,
+            lat_lim=lat_lim,
+            lon_lim=lon_lim,
+            fmt=fmt,
+        )
+        self.output_folder = os.path.join(
+            Path(path).absolute(), "chirps", "precipitation"
+        )
 
         # make directory if it not exists
         if not os.path.exists(self.output_folder):
             os.makedirs(self.output_folder)
 
-        # check time variables
-        if start == "":
-            self.start = pd.Timestamp("1981-01-01")
+    def check_input_dates(
+        self, start: str, end: str, temporal_resolution: str, fmt: str
+    ):
+        """check validity of input dates.
+
+        Parameters
+        ----------
+        temporal_resolution: (str, optional)
+            [description]. Defaults to 'daily'.
+        start: (str, optional)
+            [description]. Defaults to ''.
+        end: (str, optional)
+            [description]. Defaults to ''.
+        fmt: (str, optional)
+            [description]. Defaults to "%Y-%m-%d".
+        """
+        # check temporal_resolution variables
+        if start is None:
+            self.start = pd.Timestamp(self.start_date)
         else:
             self.start = dt.datetime.strptime(start, fmt)
 
-        if end == "":
-            self.end = pd.Timestamp("Now")
+        if end is None:
+            self.end = pd.Timestamp(self.end_date)
         else:
             self.end = dt.datetime.strptime(end, fmt)
-        # Create days
-        self.Dates = pd.date_range(self.start, self.end, freq=self.time_freq)
 
+        # Define timestep for the timedates
+        if temporal_resolution.lower() == "daily":
+            self.time_freq = "D"
+            # self.output_folder = os.path.join(path, "precipitation", "chirps", "daily")
+        elif temporal_resolution.lower() == "monthly":
+            self.time_freq = "MS"
+            # self.output_folder = os.path.join(
+            #     path, "Precipitation", "CHIRPS", "Monthly"
+            # )
+        else:
+            raise KeyError("The input temporal_resolution interval is not supported")
+
+        # Create days
+        self.dates = pd.date_range(self.start, self.end, freq=self.time_freq)
+
+    def initialize(self):
+        """Initialize FTP server."""
+        print("FTP server datasources does not need server initialization")
+        pass
+
+    def create_grid(self, lat_lim: list, lon_lim: list):
+        """Create_grid.
+
+            create grid from the lat/lon boundaries
+
+        Parameters
+        ----------
+        lat_lim: []
+            latitude boundaries
+        lon_lim: []
+            longitude boundaries
+        """
+        self.lat_lim = []
+        self.lon_lim = []
         # Check space variables
-        if lat_lim[0] < -50 or lat_lim[1] > 50:
+        # -50 , 50
+        if lat_lim[0] < self.lat_bondaries[0] or lat_lim[1] > self.lat_bondaries[1]:
             print(
                 "Latitude above 50N or below 50S is not possible."
                 " Value set to maximum"
             )
-            self.lat_lim[0] = np.max(lat_lim[0], -50)
-            self.lat_lim[1] = np.min(lon_lim[1], 50)
-        if lon_lim[0] < -180 or lon_lim[1] > 180:
+            self.lat_lim[0] = np.max(lat_lim[0], self.lat_bondaries[0])
+            self.lat_lim[1] = np.min(lon_lim[1], self.lat_bondaries[1])
+        # -180, 180
+        if lon_lim[0] < self.lon_boundaries[0] or lon_lim[1] > self.lon_boundaries[1]:
             print(
                 "Longitude must be between 180E and 180W."
                 " Now value is set to maximum"
             )
-            self.lon_lim[0] = np.max(lat_lim[0], -180)
-            self.lon_lim[1] = np.min(lon_lim[1], 180)
+            self.lon_lim[0] = np.max(lat_lim[0], self.lon_boundaries[0])
+            self.lon_lim[1] = np.min(lon_lim[1], self.lon_boundaries[1])
         else:
             self.lat_lim = lat_lim
             self.lon_lim = lon_lim
+
         # Define IDs
         self.yID = 2000 - np.int16(
             np.array(
@@ -107,10 +169,10 @@ class CHIRPS:
             )
         )
 
-    def Download(self, progress_bar: bool = True, cores=None):
+    def download(self, progress_bar: bool = True, cores=None, *args, **kwargs):
         """Download.
 
-        Download method downloads CHIRPS data
+            downloads CHIRPS data
 
         Parameters
         ----------
@@ -128,7 +190,7 @@ class CHIRPS:
         # Pass variables to parallel function and run
         args = [
             self.output_folder,
-            self.time,
+            self.temporal_resolution,
             self.xID,
             self.yID,
             self.lon_lim,
@@ -138,7 +200,7 @@ class CHIRPS:
         if not cores:
             # Create Waitbar
             if progress_bar:
-                total_amount = len(self.Dates)
+                total_amount = len(self.dates)
                 amount = 0
                 print_progress_bar(
                     amount,
@@ -148,8 +210,9 @@ class CHIRPS:
                     length=50,
                 )
 
-            for Date in self.Dates:
-                CHIRPS.RetrieveData(Date, args)
+            for date in self.dates:
+                self.API(date, args)
+
                 if progress_bar:
                     amount = amount + 1
                     print_progress_bar(
@@ -162,23 +225,73 @@ class CHIRPS:
             results = True
         else:
             results = Parallel(n_jobs=cores)(
-                delayed(CHIRPS.RetrieveData)(Date, args) for Date in self.Dates
+                delayed(self.API)(date, args) for date in self.dates
             )
         return results
 
+    def API(self, date, args):
+        """form the request url abd trigger the request.
+
+        Parameters
+        ----------
+        date:
+
+        args: [list]
+        """
+        [output_folder, temp_resolution, xID, yID, lon_lim, latlim] = args
+
+        # Define FTP path to directory
+        if temp_resolution.lower() == "daily":
+            pathFTP = f"pub/org/chg/products/CHIRPS-2.0/global_daily/tifs/p05/{date.strftime('%Y')}/"
+        elif temp_resolution == "monthly":
+            pathFTP = "pub/org/chg/products/CHIRPS-2.0/global_monthly/tifs/"
+        else:
+            raise KeyError("The input temporal_resolution interval is not supported")
+
+        # create all the input name (filename) and output (outfilename, filetif, DiFileEnd) names
+        if temp_resolution.lower() == "daily":
+            filename = f"{self.globe_fname}.{date.strftime('%Y')}.{date.strftime('%m')}.{date.strftime('%d')}.tif.gz"
+            outfilename = os.path.join(
+                output_folder,
+                f"{self.globe_fname}.{date.strftime('%Y')}.{date.strftime('%m')}.{date.strftime('%d')}.tif",
+            )
+            DirFileEnd = os.path.join(
+                output_folder,
+                f"{self.clipped_fname}_mm-day-1_daily_{date.strftime('%Y')}.{date.strftime('%m')}.{date.strftime('%d')}.tif",
+            )
+        elif temp_resolution == "monthly":
+            filename = (
+                f"{self.globe_fname}.{date.strftime('%Y')}.{date.strftime('%m')}.tif.gz"
+            )
+            outfilename = os.path.join(
+                output_folder,
+                f"{self.globe_fname}.{date.strftime('%Y')}.{date.strftime('%m')}.tif",
+            )
+            DirFileEnd = os.path.join(
+                output_folder,
+                f"{self.clipped_fname}_mm-month-1_monthly_{date.strftime('%Y')}.{date.strftime('%m')}.{date.strftime('%d')}.tif",
+            )
+        else:
+            raise KeyError("The input temporal_resolution interval is not supported")
+
+        self.callAPI(pathFTP, output_folder, filename)
+        self.post_download(
+            output_folder, filename, lon_lim, latlim, xID, yID, outfilename, DirFileEnd
+        )
+
     @staticmethod
-    def RetrieveData(Date, args):
-        """RetrieveData.
+    def callAPI(pathFTP: str, output_folder: str, filename: str):
+        """send the request to the server.
 
         RetrieveData method retrieves CHIRPS data for a given date from the
         https://data.chc.ucsb.edu/
 
         Parameters
         ----------
-        Date : TYPE
-            DESCRIPTION.
-        args : TYPE
-            A list of parameters defined in the DownloadData function.
+        filename
+        output_folder
+        pathFTP
+
 
         Raises
         ------
@@ -190,25 +303,8 @@ class CHIRPS:
         bool
             DESCRIPTION.
         """
-        # Argument
-        [output_folder, TimeCase, xID, yID, lon_lim, latlim] = args
-
-        # open ftp server
-        # ftp = FTP("chg-ftpout.geog.ucsb.edu", "", "")
-        ftp = FTP("data.chc.ucsb.edu")
+        ftp = FTP(CHIRPS.api_url)
         ftp.login()
-
-        # Define FTP path to directory
-        if TimeCase.lower() == "daily":
-            pathFTP = (
-                "pub/org/chg/products/CHIRPS-2.0/global_daily/tifs/p05/%s/"
-                % Date.strftime("%Y")
-            )
-        elif TimeCase == "monthly":
-            pathFTP = "pub/org/chg/products/CHIRPS-2.0/global_monthly/tifs/"
-        else:
-            raise KeyError("The input time interval is not supported")
-
         # find the document name in this directory
         ftp.cwd(pathFTP)
         listing = []
@@ -216,47 +312,39 @@ class CHIRPS:
         # read all the file names in the directory
         ftp.retrlines("LIST", listing.append)
 
-        # create all the input name (filename) and output (outfilename, filetif, DiFileEnd) names
-        if TimeCase.lower() == "daily":
-            filename = "chirps-v2.0.%s.%02s.%02s.tif.gz" % (
-                Date.strftime("%Y"),
-                Date.strftime("%m"),
-                Date.strftime("%d"),
-            )
-            outfilename = os.path.join(
-                output_folder,
-                "chirps-v2.0.%s.%02s.%02s.tif"
-                % (Date.strftime("%Y"), Date.strftime("%m"), Date.strftime("%d")),
-            )
-            DirFileEnd = os.path.join(
-                output_folder,
-                "P_CHIRPS.v2.0_mm-day-1_daily_%s.%02s.%02s.tif"
-                % (Date.strftime("%Y"), Date.strftime("%m"), Date.strftime("%d")),
-            )
-        elif TimeCase == "monthly":
-            filename = "chirps-v2.0.%s.%02s.tif.gz" % (
-                Date.strftime("%Y"),
-                Date.strftime("%m"),
-            )
-            outfilename = os.path.join(
-                output_folder,
-                "chirps-v2.0.%s.%02s.tif" % (Date.strftime("%Y"), Date.strftime("%m")),
-            )
-            DirFileEnd = os.path.join(
-                output_folder,
-                "P_CHIRPS.v2.0_mm-month-1_monthly_%s.%02s.%02s.tif"
-                % (Date.strftime("%Y"), Date.strftime("%m"), Date.strftime("%d")),
-            )
-        else:
-            raise KeyError("The input time interval is not supported")
-
         # download the global rainfall file
-        try:
-            local_filename = os.path.join(output_folder, filename)
-            lf = open(local_filename, "wb")
-            ftp.retrbinary("RETR " + filename, lf.write, 8192)
-            lf.close()
+        local_filename = os.path.join(output_folder, filename)
+        lf = open(local_filename, "wb")
+        ftp.retrbinary("RETR " + filename, lf.write, 8192)
+        lf.close()
 
+    def post_download(
+        self,
+        output_folder,
+        filename,
+        lon_lim,
+        latlim,
+        xID,
+        yID,
+        outfilename,
+        DirFileEnd,
+    ):
+        """clip the downloaded data to the extent we want.
+
+        Parameters
+        ----------
+        output_folder: [str]
+            directory where files will be saved
+        filename: [str]
+            file name
+        lon_lim: [list]
+        latlim: [list]
+        xID: [list]
+        yID: [list]
+        outfilename: [str]
+        DirFileEnd: [str]
+        """
+        try:
             # unzip the file
             zip_filename = os.path.join(output_folder, filename)
             extractFromGZ(zip_filename, outfilename, delete=True)
@@ -289,7 +377,7 @@ class CHIRPS:
             )
         return True
 
-    def ListAttributes(self):
+    def listAttributes(self):
         """Print Attributes List."""
 
         print("\n")
@@ -303,3 +391,26 @@ class CHIRPS:
                 print(str(key) + " : " + repr(self.__dict__[key]))
 
         print("\n")
+
+
+class Catalog(AbstractCatalog):
+    """CHIRPS data catalog."""
+
+    def __init__(self):
+        self.catalog = self.get_catalog()
+
+    def get_catalog(self):
+        """return the catalog."""
+        return {
+            "Precipitation": {
+                "descriptions": "rainfall [mm/temporal_resolution step]",
+                "units": "mm/temporal_resolution step",
+                "temporal resolution": ["daily", "monthly"],
+                "file name": "rainfall",
+                "var_name": "R",
+            }
+        }
+
+    def get_variable(self, var_name):
+        """get the details of a specific variable."""
+        return self.catalog.get(var_name)