+++ tests/mo/mswep.py
@@ -1,0 +1,2 @@
+# class MSWEP():
+"""http://www.gloh2o.org/mswx/"""